+++ src/ui/action_handler.py
@@ -1,9 +1,12 @@
-from PyQt6.QtWidgets import QDialog, QFileDialog
+from PyQt6.QtWidgets import QDialog, QFileDialog, QMessageBox
 
+from operations.created_to_mod import created_to_mod_for_folder_path_operation, \
+    created_to_mod_for_list_of_files_operation
 from operations.flat_to_tree import flat_to_tree_operation
 from operations.heic_to_jpg import heic_to_jpg_operation
 from operations.takeout import takeout_operation
 from ui.browser import Browser
+from ui.designer.created_to_mod import Ui_created_to_mod
 from ui.designer.flat_to_tree import Ui_flat_to_tree
 from ui.designer.heic_to_jpg import Ui_heic_to_jpg
 from ui.designer.takeout import Ui_takeout
@@ -22,6 +25,7 @@ class ActionHandler:
         self.action_buttons['takeout'].pressed.connect(self.handle_takeout)
         self.action_buttons['heic_to_jpg'].released.connect(self.handle_heic_to_jpg)
         self.action_buttons['flat_to_tree'].released.connect(self.handle_flat_to_tree)
+        self.action_buttons['created_to_mod'].released.connect(self.handle_created_to_mod)
         # self.btn_number.pressed.connect(self.numberButtonAction)
         # self.btn_rename.pressed.connect(self.renameButtonAction)
         # self.btn_resize.pressed.connect(self.resizeButtonAction)
@@ -151,6 +155,59 @@ class ActionHandler:
 
         # create the dialog and show it,
         dlg = FlatToTreeDialog(self.folder_select.folder_edit.text(), self.parent)
+        dlg.exec()
+        # perform follow-up actions after closing
+        if dlg.go_to_output:
+            self.folder_select.force_set_directory(dlg.edit_folder_path.text())
+        else:
+            self.folder_select.force_refresh()
+
+    def handle_created_to_mod(self):
+        # define a wrapper class for the dialog
+        class CreatedToModDialog(QDialog, Ui_created_to_mod):
+            def __init__(self, initial_folder: str, current_selection: list[str], parent=None):
+                QDialog.__init__(self, parent)
+                self.setupUi(self)
+                self.edit_folder_path.setText(initial_folder)
+                self.go_to_output = False
+                self.current_selection = current_selection
+
+                # slots
+                self.btn_perform_action.clicked.connect(self.start_created_to_mod)
+                self.btn_folder_select.clicked.connect(
+                    lambda: self.edit_folder_path.setText(
+                        QFileDialog.getExistingDirectory(self, 'Select Images Folder', self.edit_folder_path.text())
+                    )
+                )
+                self.btn_close_redirect.clicked.connect(self.on_close_redirect)
+
+            def start_created_to_mod(self):
+                self.text_output.clear()
+
+                def callback(message, progress):
+                    self.progress_bar.setValue(progress)
+                    self.text_output.append(message)
+
+                # call entrypoint for folder or files depending on ui configuration
+
+                if self.rd_selected_files_only.isChecked():
+                    # user selected to process only selected files
+                    if len(self.current_selection) == 0:
+                        QMessageBox.warning(self, "No selection", "Create a selection first.")
+                        return
+                    created_to_mod_for_list_of_files_operation(self.current_selection, callback)
+                else:
+                    # user selected to process entire folder
+                    folder_path = self.edit_folder_path.text()
+                    created_to_mod_for_folder_path_operation(folder_path, callback)
+
+            def on_close_redirect(self):
+                self.go_to_output = True
+                self.accept()
+
+        # create the dialog and show it,
+        current_selection = [file_info.absoluteFilePath() for file_info in self.browser.get_selection()]
+        dlg = CreatedToModDialog(self.folder_select.folder_edit.text(), current_selection, self.parent)
         dlg.exec()
         # perform follow-up actions after closing
         if dlg.go_to_output: